files updated

+++ assignment 24/Nine_max_of_two_overloading_function.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 using namespace std;
-int max_of_two(int, int);
+// int max_of_two(int, int);
+float max_of_two(float, float);
 int main()
 {
     int no1, no2;
@@ -10,7 +11,11 @@ int main()
     cout << max_of_two(no1, no2);
     return 0;
 }
-int max_of_two(int a, int b)
+// int max_of_two(int a, int b)
+// {
+//     return a > b ? a : b;
+// }
+float max_of_two(float a, float b)
 {
     return a > b ? a : b;
 }
